Added easing functions.
On the seventh day, God said Let There Be Light!  And the light popped on and people were like Ouch, that was harsh!  On the eighth day, God said Let The Light Increase in a Linear Fashion! and people were like, Ok, that's better!  And then God became bored and on the ninth day he said Let The Light Increase in a Non-linear Fashion! and the people were like Oh, that's pretty cool!