BasicBird and BirdBase for bird rendering.
Base class for implementing birds.  Each child class has the option for a Midi Channel and Midi Note number but that might not be used.  Each look can implement BirdBase and then in the future we might look through the patterns in channel for any pattern where note = 50 and then call out to the renderBirdSinging, renderBirdIdle methods directly while passing in our own color buffer.  This allows for looks to be developed and prototyped while also allowing for whatever the future MIDI integration might be.