Provide polar angle sorted points.
For each section like upperRingFloods there is now an upperRingFloodsSorted list with the same points but in sorted order.  I need to keep the original file based order as it was exported from Rhino because that corresponds with the light positions that were imported into Unity.  The wiring will probably want to use polar angle sorted lights.  and there should be some agreement on where 0 degrees starts in physical space respective of the wiring.