Stereo option for birds.
This works on my Mac.  The audio is routed to just 2 channels instead of 6.  You will need to pip install python-osc before speaker.py will work.  Also, you need to install MidiPipe to create a MIDI port that supercollider can write to and that FL Studio will read from.  The stereobirds.lxp file is set up with MIDI enabled and routed to a Birds pattern so that when you run speaker.py 5757 1 0.4 it will trigger birds on speaker 1 at 0.4 volume.  5757 is the port supercollider is listening on.  Running scmac.sh from the sc/ directory will start up sclang (supercollider interpreter) with the file sirstereo.scd.