Bird size fudging and volume tracking bird audio interactivity.
Any directories under sc/ will be loaded as WAV synths now.  There is a new python script to trigger them.  Example: python synth.py tracs 3 1 0.02 10.0  Which means play 3.wav from the tracs subdirectory on audio channel 1 at volume 0.02 with a 10.0 scale multiplier for the amplitude value reported back to LX studio.  The last value is to make it easier to see some effect when the volume is relatively low and allow for some visual tuning later.  The new subdirectory synths stuff is in sirsasanasynth.scd or run with sirsynth.bat.