Pixlite output mapping
I just picked some random defaults.  For a particular group of flood lights, the lights should always be wired in counter-clockwise order.  For example, cg.3 means canopy group 3, so the two lights on the third canopy.  The first light in wire-order should be the light that is more clockwise and the second light should be the light that is more counter-clockwise.  Both canopy floods and ground floods are grouped for wire/output mapping purposes just to simplify the amount of data that needs to be input.  Also, in order to use the Unity output now, you should create a file called 'unity' in the project directory.  If that file exists, Sirsasana will choose to output ArtNet to Unity intead.