Basic bird song generator.
Retriggers every 30 seconds.  There currently isn't a great way to figure out the individual bird synth lengths from LX Studio so right now we just wait for 30 seconds and retrigger.  It is possible using PlayBuf's but we might need to migrate everything to streaming from disk with DiskIn.  In that scenario it is difficult to detect the file length.