OUTPUT.md are wiring instructions. Also implemented offsets for wiring.
Offsets allow you to adjust which light LX Studio thinks is the first light in terms of data wiring.  Everything is sorted by polar angle (azimuth) around the installation so for each area of the installation, it thinks the first light is closest to azimuth = 0.  If the wiring is not consistent, it is possible to shift which light LX Studio thinks is the first light with respect  to the data wiring by adjusting the offset for that particular light group.  For groups of lights like the ground floods, the offset refers to the group number.  For example there are 3 groups of 3 for ground floods.  Within a group of 3 the floods should be wired in counterclockwise order.  The offset is an offset for groups so the valid values are 0, 1, 2 and it which adjust the group of lights that LX Studio thinks are first on the data wire.