OutputTest pattern.
The OutputTest pattern allows for specification of the group number within a section for the ground floods and canopy sections.  This can be useful for testing the wiring.  It is also a compound parameter so it can be used with a linear envelop to cycle the lights around the sculpture.