Tyson Audio.
I like big buffers and I cannot lie.  I think heavy OSC activity was starving the streaming from disk and crashing the audio so I increased the bufferSize for the synth buffers.  The bufferSize has been bumped to 128k and the OSC update frequency reduced to 16 times per second.  Also we only send OSC messages if the amplitude is greater than 0.01 which should reduce OSC contention. The wave files will need to be managed out-of-band of github.