Control signal remapping.
Allows for a variable number of birds.  sirsasanasynthd.scd has a birdToBus array that controls how bird wav files are mapped to control bus numbers.  It currently maps 2 consecutive bird WAVs to the same control bus so we end up with 6 active control busses.  In LX Studio, there is a ctrl sig mapping panel that then maps the first 6 control signals to the 12 birds in the model whereas 1 signal is mapped to 2 birds in the model.  The model in LX will always have 12 birds and the bird numbers determine their position in the model.  If there only exists 6 birds to be wired up, then the wired birds should be b.1, b.3, b.5, ..., b.11 which correspond to the first bird on each speaker in physical space.  Technically, the ctrl sig mapping could just be 1, 3, 5, 7, 9, 11 instead of replicating the signal to multiple birds.  I just replicated the signal in the case that the output mapping is not correct although I've also removed b.2, b.4, etc from the default ArtNet output mapping.