Updated show file.  Fixes cut-out by Bird patterns transitioning.
While bird patterns transition they were losing their alpha channel which was causing the rest of the installation to black out.  I fixed it by having each bird pattern initialize all points to full black and then moving the bird channel to the first channel in the project.  The sculpture panel now uses the MultiMask effect to write to everything except the birds.